FIX: Development mode support
Details:

We had a lot of problem in the past with development mode.
This change plans to make it stable once and for all.

1. Fix `mode`

The `mode` property on the configuration object would fail to obtain
mode if the NODE_ENV variable hadn't been set.
Now, it uses `production` property from the bundler to determine the
correct mode.

2. Remove `updateNodeModulesPath`

The function was redundant since the node_modules path in the map/style
files is relative to the source directory, not the output directory.

3. Add `glob` library

To help with the path calculation, I've introduced the `glob` library
that supports bash-like globbing.
The main goal was to refactor updateDependecies() and updateDependants()
functions to support handling paths in development mode.