FIX: Bug with source map links
When the map file is processed in the development mode (user specified ".map": "folder")
it will wrongly update it's path in the dependant file (it collides with
it)
Example wrong path that we get -> i.e. ../maps/maps/style.2kj52.css.map
To prevent this an `if` statement was added to prevent map file from
updating it's path in the dependant file.