FIX: buildDependencyAssets() stack overflow
On bigger projects (800+ files) buildDependecyAssets() function would
cause plugin to crash. It would eat all memory because of it's recursive
design pattern.

I've refactored it in 2 separate functions:
1) buildDepAssets()
2) buildJSAssetDependecies()

The first one is for entry files imports and the second is for JS dist
file imports. This approach avoids stack overflow.

Minor Changes:
 - DependencyGraph test that tests if dependecies are connected properly
to their dependants

Resolves #8